backend/requirements.txt
fastapi
uvicorn
chromadb
anthropic
python-dotenv
pydantic
python-jose[cryptography]